sort areas for miou

+++ torchao/_models/sam2/modeling/sam2_base.py
@@ -467,14 +467,15 @@ class SAM2Base(torch.nn.Module):
     def forward_image(self, img_batch: torch.Tensor):
         """Get the image feature on the input batch."""
         backbone_out = self.image_encoder(img_batch)
+        # import pdb; pdb.set_trace()
         if self.use_high_res_features_in_sam:
             # precompute projected level 0 and level 1 features in SAM decoder
             # to avoid running it again on every SAM click
             backbone_out["backbone_fpn"][0] = self.sam_mask_decoder.conv_s0(
-                backbone_out["backbone_fpn"][0]
+                backbone_out["backbone_fpn"][0].to(torch.float32)
             )
             backbone_out["backbone_fpn"][1] = self.sam_mask_decoder.conv_s1(
-                backbone_out["backbone_fpn"][1]
+                backbone_out["backbone_fpn"][1].to(torch.float32)
             )
         return backbone_out
